Bug : tentative insertion vente toujours bugée malgré les changements

+++ dvdstore/src/main/java/com/simplon/dvdstore/dvdstore/services/DvdModelServiceSales.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class DvdModelServiceSales {
-    private int FKUsers;
-    private int FKMovies;
-    private float quantityOfSales;
+    private Long FKUsers;
+    private Long FKMovies;
+    private Long quantityOfSales;
 }
